requirements.txt
# Python 3.11.9
python-dotenv==1.0.1
openai==1.58.1
opencv-python==4.10.0.84
streamlit==1.41.1
onnxruntime==1.20.1
